Cleaned and improved card selectors
