+++ stegraca-valise/package.json
@@ -11,8 +11,8 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@yoieh/use-holderjs": "^0.0.11",
     "bootstrap": "^5.2.3",
+    "extract-colors": "^4.0.2",
     "gh-pages": "^5.0.0",
     "holderjs": "^2.9.9",
     "react": "^18.2.0",
