+++ stegraca-valise/src/components/project/Project.css
@@ -16,9 +16,9 @@
     transition-property: width;
     transition-duration: 0.5s;
     flex-direction: row;
-    margin: 10px;
+    margin: 5px;
     justify-content: flex-start;
-
+    align-items: center;
 }
 
 .navs {
@@ -64,6 +64,9 @@
 .project-details {
     grid-area: project-details;
     position:inherit;
+    padding: 10px 10px 10px 10px ;
+    margin-left: 10px;
+    background: rgba(0, 0, 0, 0.25);
 }
 
 .card-image {
@@ -83,7 +86,7 @@
     /* box-shadow: 0px 0px 1px #000000; */
     font-size: larger;
     cursor: pointer;
-    padding: 10px;
+    padding: 10px 10px 10px 10px;
     transition: all ease-in-out 0.25s;
 }
 
@@ -167,11 +170,24 @@
 }
 
 .open {
-    width: 25%
+    width: 25%;
 }
 
 .closed {
     width: 100%
+}
+
+.card-open {
+    padding: 5px;
+}
+
+.card-selected {
+    background: rgba(0, 0, 0, 0.25);
+    width: 105%;
+}
+
+.card-closed {
+
 }
 
 .hd-off {
